# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Insta Deploy API
  version: 0.2.0
  description: |
    Deploy Docker. Get a URL.

    **Authentication.** Accounts and sessions are managed by Better Auth in the dashboard
    (`POST /api/auth/sign-up/email`, `POST /api/auth/sign-in/email` on the dashboard's URL, port 3000).
    Sign-in returns a `token`; send it to this API as `Authorization: Bearer <token>`.
    The dashboard itself uses the session cookie.

    The **agent API** (`/api/agent/*`) is used by the Insta Deploy agent with its own token.
servers:
  - url: /
...
components:
  securitySchemes:
    session:
      type: http
      scheme: bearer
      description: Better Auth session token
    agentToken:
      type: http
      scheme: bearer
      description: Agent token (idt_...)
...
security:
  - session: []
paths:
...
  /api/analyze/git:
    post:
      tags: [Analysis]
      summary: Resolve a branch and let the agent look inside the repository
      requestBody:
        content: { application/json: { schema: { type: object, properties: { git_url: { type: string }, git_branch: { type: string }, agent_id: { type: string } } } } }
      responses: { "200": { description: OK } }
  /api/github:
    get:
      tags: [GitHub]
      summary: Your GitHub App, where it's installed and the repositories it can read
      responses: { "200": { description: OK } }
    put:
      tags: [GitHub]
      summary: Connect or update your GitHub App (checked with GitHub before saving)
      requestBody:
        content: { application/json: { schema: { type: object, properties: { app_id: { type: string, example: "123456" }, private_key: { type: string, description: PEM; empty keeps the saved key }, webhook_secret: { type: string, description: Empty keeps the saved secret } } } } }
      responses: { "200": { description: OK }, "400": { description: Invalid input or missing Contents permission }, "502": { description: GitHub rejected the credentials } }
    delete: { tags: [GitHub], summary: Disconnect your GitHub App, responses: { "204": { description: Disconnected } } }
  /api/github/webhook:
    post:
      tags: [GitHub]
      security: []
      summary: GitHub App webhook (signed with the app's webhook secret). A push redeploys auto-deploy deployments of that repository and branch.
      responses: { "200": { description: Ping }, "202": { description: Accepted }, "401": { description: Invalid signature } }
...